+++ src/Engine/PdfParser/CrossReference/ReaderInterface.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * This file is part of FPDI
+ *
+ * @package   YnbAgency\Fpdi\Engine
+ * @copyright Copyright (c) 2026 Setasign GmbH & Co. KG (https://www.setasign.com)
+ * @license   http://opensource.org/licenses/mit-license The MIT License
+ */
+
+namespace YnbAgency\Fpdi\Engine\PdfParser\CrossReference;
+
+use YnbAgency\Fpdi\Engine\PdfParser\Type\PdfDictionary;
+
+/**
+ * ReaderInterface for cross-reference readers.
+ */
+interface ReaderInterface
+{
+    /**
+     * Get an offset by an object number.
+     *
+     * @param int $objectNumber
+     * @return int|bool False if the offset was not found.
+     */
+    public function getOffsetFor($objectNumber);
+
+    /**
+     * Get the trailer related to this cross reference.
+     *
+     * @return PdfDictionary
+     */
+    public function getTrailer();
+}