+++ src/Engine/PdfParser/Filter/Ascii85Exception.php
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * This file is part of FPDI
+ *
+ * @package   YnbAgency\Fpdi\Engine
+ * @copyright Copyright (c) 2026 Setasign GmbH & Co. KG (https://www.setasign.com)
+ * @license   http://opensource.org/licenses/mit-license The MIT License
+ */
+
+namespace YnbAgency\Fpdi\Engine\PdfParser\Filter;
+
+/**
+ * Exception for Ascii85 filter class
+ */
+class Ascii85Exception extends FilterException
+{
+    /**
+     * @var integer
+     */
+    const ILLEGAL_CHAR_FOUND = 0x0301;
+
+    /**
+     * @var integer
+     */
+    const ILLEGAL_LENGTH = 0x0302;
+}