+++ src/Engine/PdfParser/Filter/FilterInterface.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of FPDI
+ *
+ * @package   YnbAgency\Fpdi\Engine
+ * @copyright Copyright (c) 2026 Setasign GmbH & Co. KG (https://www.setasign.com)
+ * @license   http://opensource.org/licenses/mit-license The MIT License
+ */
+
+namespace YnbAgency\Fpdi\Engine\PdfParser\Filter;
+
+/**
+ * Interface for filters
+ */
+interface FilterInterface
+{
+    /**
+     * Decode a string.
+     *
+     * @param string $data The input string
+     * @return string
+     */
+    public function decode($data);
+}