+++ src/Engine/PdfParser/Type/PdfTypeException.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * This file is part of FPDI
+ *
+ * @package   YnbAgency\Fpdi\Engine
+ * @copyright Copyright (c) 2026 Setasign GmbH & Co. KG (https://www.setasign.com)
+ * @license   http://opensource.org/licenses/mit-license The MIT License
+ */
+
+namespace YnbAgency\Fpdi\Engine\PdfParser\Type;
+
+use YnbAgency\Fpdi\Engine\PdfParser\PdfParserException;
+
+/**
+ * Exception class for pdf type classes
+ */
+class PdfTypeException extends PdfParserException
+{
+    /**
+     * @var int
+     */
+    const NO_NEWLINE_AFTER_STREAM_KEYWORD = 0x0601;
+}