+++ src/Engine/PdfReader/PdfReaderException.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * This file is part of FPDI
+ *
+ * @package   YnbAgency\Fpdi\Engine
+ * @copyright Copyright (c) 2026 Setasign GmbH & Co. KG (https://www.setasign.com)
+ * @license   http://opensource.org/licenses/mit-license The MIT License
+ */
+
+namespace YnbAgency\Fpdi\Engine\PdfReader;
+
+use YnbAgency\Fpdi\Engine\FpdiException;
+
+/**
+ * Exception for the pdf reader class
+ */
+class PdfReaderException extends FpdiException
+{
+    /**
+     * @var int
+     */
+    const KIDS_EMPTY = 0x0101;
+
+    /**
+     * @var int
+     */
+    const UNEXPECTED_DATA_TYPE = 0x0102;
+
+    /**
+     * @var int
+     */
+    const MISSING_DATA = 0x0103;
+}